Separate files etc.
1.) Separate file
Parser.scala -> Parser.scala, Scale.scala

2.) Change arrangement
Launcher object in com.yamanogusha.scale.parser -> Launcher object in com.yamanogusha.scale

2.) Change Process
Launcher object @ Scale.scala
Read preparation file
